#1406 [VA-IIR] - V3 Profile Upgrade: Feature flag/unused code cleanup (#2055)
