+++ cd/application-deployment/dev/vaec-api-task-definition.json
@@ -222,10 +222,6 @@
         {
           "name": "COMP_AND_PEN_DYNAMODB_NAME",
           "value": "dev-bip-payment-notification-table"
-        },
-        {
-          "name": "VA_PROFILE_V3_COMBINE_CONTACT_INFO_AND_PERMISSIONS_LOOKUP",
-          "value": "True"
         }
       ],
       "secrets": [
