+++ cd/application-deployment/perf/vaec-api-task-definition.json
@@ -198,10 +198,6 @@
         {
           "name": "COMP_AND_PEN_DYNAMODB_NAME",
           "value": "perf-bip-payment-notification-table"
-        },
-        {
-          "name": "VA_PROFILE_V3_COMBINE_CONTACT_INFO_AND_PERMISSIONS_LOOKUP",
-          "value": "True"
         }
       ],
       "secrets": [
