+++ cd/application-deployment/perf/vaec-celery-task-definition.json
@@ -188,14 +188,6 @@
         {
           "name": "COMP_AND_PEN_PERF_TO_NUMBER",
           "value": "+14254147755"
-        },
-        {
-          "name": "VA_PROFILE_V3_COMBINE_CONTACT_INFO_AND_PERMISSIONS_LOOKUP",
-          "value": "True"
-        },
-        {
-          "name": "VA_PROFILE_V3_IDENTIFY_MOBILE_TELEPHONE_NUMBERS",
-          "value": "True"
         }
       ],
       "secrets": [
