+++ cd/application-deployment/prod/vaec-api-task-definition.json
@@ -202,10 +202,6 @@
         {
           "name": "COMP_AND_PEN_DYNAMODB_NAME",
           "value": "prod-bip-payment-notification-table"
-        },
-        {
-          "name": "VA_PROFILE_V3_COMBINE_CONTACT_INFO_AND_PERMISSIONS_LOOKUP",
-          "value": "True"
         }
       ],
       "secrets": [
