+++ cd/application-deployment/staging/vaec-api-task-definition.json
@@ -218,10 +218,6 @@
         {
           "name": "COMP_AND_PEN_DYNAMODB_NAME",
           "value": "staging-bip-payment-notification-table"
-        },
-        {
-          "name": "VA_PROFILE_V3_COMBINE_CONTACT_INFO_AND_PERMISSIONS_LOOKUP",
-          "value": "True"
         }
       ],
       "secrets": [
