+++ cd/application-deployment/staging/vaec-celery-task-definition.json
@@ -208,14 +208,6 @@
         {
           "name": "COMP_AND_PEN_DYNAMODB_NAME",
           "value": "staging-bip-payment-notification-table"
-        },
-        {
-          "name": "VA_PROFILE_V3_COMBINE_CONTACT_INFO_AND_PERMISSIONS_LOOKUP",
-          "value": "True"
-        },
-        {
-          "name": "VA_PROFILE_V3_IDENTIFY_MOBILE_TELEPHONE_NUMBERS",
-          "value": "True"
         }
       ],
       "secrets": [
